xfz1/requirements.txt
asgiref==3.2.3
bleach==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
Django==3.0.2
django-debug-toolbar==2.2
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.1.2
PyMySQL==0.9.3
python-memcached==1.59
pytz==2019.3
qiniu==7.2.8
redis==3.5.0
requests==2.23.0
shortuuid==1.0.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.0
